Moved some files arround and minor fixes
